Refactor E2E tests: update discount tests for blocks and classic checkouts, enhance checkout helpers, and remove obsolete test files
