+++ package.json
@@ -26,7 +26,7 @@
     "i18n:json": "$(which wp) i18n make-json languages --no-purge",
     "packages-update": "wp-scripts packages-update",
     "test": "wp-scripts test-unit-js --testPathPattern=tests/js",
-    "test:e2e": "node tests/e2e/discount-3ds-checkout.spec.js",
+    "test:e2e": "node tests/e2e/discount-3ds-classic.spec.js && node tests/e2e/discount-3ds-blocks.spec.js",
     "check-engines": "wp-scripts check-engines",
     "zip": "rm deploy.zip || true && npm i && npm run build && composer install && zip -r deploy.zip ./ -x '*.git*'  -x '*.idea*' -x '*.github*' -x 'node_modules*' -x 'tests*' -x '.claude*'"
   },
